ref(ngrx-ducks): move typings and helpers to own directories
