+++ ngrx/ducks/src/lib/actions/empty-action.ts
@@ -1,0 +1,5 @@
+export function emptyAction<T extends new () => InstanceType<T>>(type: string) {
+  return {
+    type
+  };
+}
