+++ ngrx/ducks/src/lib/actions/loaded-action.ts
@@ -1,0 +1,6 @@
+export function loadedAction(type: string, payload: unknown) {
+  return {
+    type,
+    payload
+  };
+}
