+++ ngrx/ducks/src/lib/ducks/extract-wired-action.ts
@@ -1,5 +1,6 @@
 import { MissingActionDecoratorError, throwIf } from '../errors';
 import { ClassWithActionAnnotations } from '../typings';
+
 export function extractWiredAction<T>(
   instance: ClassWithActionAnnotations<T>,
   method: string
