+++ ngrx/ducks/src/lib/typings/action-handler-with-payload.ts
@@ -1,0 +1,4 @@
+export type ActionHandlerWithPayload<TSlice, TPayload> = (
+  state: TSlice,
+  payload: TPayload
+) => TSlice;
