+++ ngrx/ducks/src/lib/typings/action-handler-without-payload.ts
@@ -1,0 +1,1 @@
+export type ActionHandlerWithoutPayload<TSlice> = (state: TSlice) => TSlice;
