+++ ngrx/ducks/src/lib/typings/duck-service.ts
@@ -1,0 +1,2 @@
+import { DuckActionDispatcher } from './DuckActionDispatcher';
+export type DuckService<T> = { [K in keyof T]: DuckActionDispatcher<T[K]> };
