+++ ngrx/ducks/src/lib/typings/index.ts
@@ -1,4 +1,10 @@
+export * from './action-handler-with-payload';
+export * from './action-handler-without-payload';
 export * from './action-that-may-have-a-payload';
 export * from './class-with-action-annotations';
+export * from './duck-action-dispatcher';
+export * from './duck-service';
 export * from './initial-state-annotated';
+export * from './loaded-action';
+export * from './plain-action';
 export * from './reducer-function';
