+++ ngrx/ducks/src/lib/typings/loaded-action.ts
@@ -1,0 +1,8 @@
+export interface LoadedAction<TPayload> {
+  action(
+    payload: TPayload
+  ): {
+    type: string;
+    payload: TPayload;
+  };
+}
