+++ ngrx/ducks/src/lib/typings/plain-action.ts
@@ -1,0 +1,5 @@
+export interface PlainAction {
+  action(): {
+    type: string;
+  };
+}
